In addition to menuentry, also don't measure submenu and [...] commands.
Both menuentry and submenu commands end up being very long because all
the contents between the braces is also part of the command arguments.
The conditional between square brackets is also not very useful to
measure since a command cannot be executed between two square brackets.
This is from skimming through the code in grub-core/commands/test.c.

The commands inside menuentry/submenu are all measured anyway so there
is no harm in not measuring them.